requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.7.0
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
av==11.0.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.30
botocore==1.34.30
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.45
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.1
curl_cffi==0.6.0b9
cycler==0.12.1
dataclasses-json==0.6.3
decorator==4.4.2
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dm-tree==0.1.8
dnspython==2.5.0
docx2txt==0.8
duckdb==0.9.2
duckduckgo_search==4.4.3
effdet==0.4.1
emoji==2.10.0
et-xmlfile==1.1.0
executing==2.0.1
facenet-pytorch==2.5.3
Faker==19.13.0
fastavro==1.9.3
fer==22.5.1
ffmpeg==1.4
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.16.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
ifaddr==0.2.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
iopath==0.1.10
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
lancedb==0.5.1
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-google-genai==0.0.6
langdetect==1.0.9
langsmith==0.0.84
layoutparser==0.3.4
libclang==16.0.6
llama-index==0.10.4
llama-index-agent-openai==0.1.1
llama-index-core==0.10.3
llama-index-embeddings-huggingface==0.1.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.1
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.12
Markdown2docx==0.1.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.10.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.18
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pikepdf==8.11.2
pillow==10.2.0
plantuml==0.3.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.1
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.7
pylibsrtp==0.10.0
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pyOpenSSL==24.0.0
pypandoc==1.12
pyparsing==3.1.1
pypdf==4.0.1
pypdfium2==4.26.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.4
pyvis==0.3.2
PyYAML==6.0.1
rapidfuzz==3.6.1
ratelimiter==1.2.0.post0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scipy==1.12.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
streamlit==1.31.0
streamlit-antd-components==0.3.2
streamlit-webrtc==0.47.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
termcolor==2.4.0
tiktoken==0.6.0
timm==0.9.12
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
twilio==8.13.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.3
unstructured-client==0.16.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
youtube-search==2.1.2
zipp==3.17.0
